+++ CHANGELOG.md
@@ -7,7 +7,7 @@ title: Changelog
 ## Unreleased (????-??-??)
 
 * New components:
-  * `<cc-grafana-info>`
+  * `<cc-grafana-info>` (with smart definition)
 
 ## 7.4.0 (2021-10-14)